+++ src/main/java/com/application/close/links/service/BridgeExecutorService.java
@@ -23,4 +23,6 @@ public interface BridgeExecutorService {
 
 	void deleteById(int executerId);
 
+	void reloadCache();
+
 }